New LOADER version

+++ picopad-loader/tools/patch_memmap.py
@@ -6,9 +6,17 @@ Import("env")
 env.Replace(PROGNAME="%s" % env.GetProjectOption("name"))
 
 FRAMEWORK_DIR = env.PioPlatform().get_package_dir("framework-arduinopico")
-original_file = join(FRAMEWORK_DIR, "lib", "memmap_default.ld")
-patched_file = join("patches", "memmap_default_bootloader.ld")
+patchflag_path = join(FRAMEWORK_DIR, ".patching-done")
 
-assert isfile(patched_file)
+# patch file only if we didn't do it before
+if not isfile(patchflag_path):
+    original_file = join(FRAMEWORK_DIR, "lib", "memmap_default.ld")
+    patched_file = join("patches", "memmap_default.ld")
 
-shutil.copy2(patched_file, original_file)
+    assert isfile(patched_file)
+
+    shutil.copy2(patched_file, original_file)
+
+    # touch the patch flag file
+    with open(patchflag_path, "w") as fp:
+        fp.write("")
